+++ src/components/ui/MicVolumeIndicator/index.tsx
@@ -1,16 +1,15 @@
 // Copyright 2020 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-import React, { FC, useRef, HTMLAttributes } from 'react';
+import React, { HTMLAttributes, Ref, forwardRef } from 'react';
 
 import Microphone from '../icons/Microphone';
 import { StyledMicVolumeIndicator } from './Styled';
 import { BaseProps } from '../Base';
 
-export interface MicVolumeIndicatorProps extends Omit<HTMLAttributes<HTMLDivElement>, 'css'>, BaseProps {
-  /* `volume` is a number from 0 to 1 that indicates the volume
-  of a  speaker's microphone in a meeting. */
-  volume: number | undefined;
+export interface MicVolumeIndicatorProps
+  extends Omit<HTMLAttributes<HTMLDivElement>, 'css'>,
+    BaseProps {
   /* `muted` indicates whether an individual is muted in a meeting. */
   muted?: boolean | undefined;
   /* `signalStrength` is a measure of an attendee's network connection on a scale of 0 to 1. 
@@ -18,19 +17,23 @@ export interface MicVolumeIndicatorProps extends Omit<HTMLAttributes<HTMLDivElem
   signalStrength: number | undefined;
 }
 
-const MicVolumeIndicator:FC<MicVolumeIndicatorProps> = (props) => {
-  const { volume, muted = false, signalStrength } = props;
-  const poorConnection = signalStrength !== undefined && signalStrength <= .5;
-  const volumeFillRef = useRef<HTMLDivElement | null>(null);
+export const MicVolumeIndicator = forwardRef(
+  (props: MicVolumeIndicatorProps, bgRef: Ref<HTMLDivElement>) => {
+    const { muted = false, signalStrength } = props;
+    const poorConnection =
+      signalStrength !== undefined && signalStrength <= 0.5;
 
-  return (
-    <StyledMicVolumeIndicator {...props} className='mic-volume-indicator'>
-      <Microphone muted={muted} poorConnection={poorConnection} className='mic-icon'/>
-      <div className='volume-fill-container'>
-        <div ref={volumeFillRef} className='volume-fill' data-testid='volume-fill'/>
-      </div>
-    </StyledMicVolumeIndicator>
-  );
-};
+    return (
+      <StyledMicVolumeIndicator {...props} className="mic-volume-indicator">
+        <Microphone
+          muted={muted}
+          className="mic-icon"
+          poorConnection={poorConnection}
+        />
+        <div ref={bgRef} className="bg-volume-fill" data-testid="volume-fill" />
+      </StyledMicVolumeIndicator>
+    );
+  }
+);
 
 export default MicVolumeIndicator;