+++ CHANGELOG.md
@@ -51,6 +51,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add introduction documentation
 - Add styles around `DeviceSelection` in demo
 - Add documentation around `DeviceSelection` and `useAudioInputActivityPreview` hook
+- Add RosterAttendee component, connected MicVolumeIndicator
 
 ### Changed
 
@@ -99,6 +100,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Remake `AudioActivityPreview` into `useLocalAudioInputActivityPreview` hook
 - Refactor styles from SDK Component `DeviceSelection` and move them to demo
 - Moved `ActivityBar` from library to demo, since it seemed more tightly related our demo
+- Stop local video, if enabled, when selecting "none" as video input source
 
 ### Removed